space

+++ exercises/EX007.py
@@ -1,4 +1,4 @@
-n1 = int(input("Primeira nota do aluno: "))
-n2 = int(input("Segunda nota do aluno: "))
+n1 = float(input("Primeira nota do aluno: "))
+n2 = float(input("Segunda nota do aluno: "))
 m = (n1 + n2) / 2
 print (f"A média entre {n1} e {n2} é igual a {m:.1f}")
